requirements.txt
fastapi
uvicorn
python-multipart
pillow
python-dotenv
mediapipe
opencv-python
openai
sqlalchemy
psycopg2-binary
slowapi 